requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.17.0
sqlparse==0.5.5
tzdata==2025.3
requests>=2.31.0
python-dotenv>=1.0.0
